+++ src/main/java/com/spring/asset_craft/entity/Product.java
@@ -13,7 +13,7 @@ public class Product {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "price")
@@ -38,7 +38,7 @@ public class Product {
     }
 
 
-    public Product(int id, String name, float price, String category, String description, float rating) {
+    public Product(Long id, String name, float price, String category, String description, float rating) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -48,10 +48,10 @@ public class Product {
     }
 
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getName() {
