+++ src/main/java/com/spring/asset_craft/repository/ProductRepository.java
@@ -10,14 +10,14 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductRepository extends JpaRepository<Product, Integer>, JpaSpecificationExecutor<Product> {
+public interface ProductRepository extends JpaRepository<Product, Long>, JpaSpecificationExecutor<Product> {
 
-    Optional<Product> findProductById(int id);
+    Optional<Product> findProductById(Long id);
 
     @Query("SELECT new com.spring.asset_craft.dto.ReviewDTO(r.comment, r.rating, u.username) " +
             "FROM Review r JOIN r.user u " +
             "WHERE r.product.id = :productId")
-    List<ReviewDTO> findProductReviews(@Param("productId") int productId);
+    List<ReviewDTO> findProductReviews(@Param("productId") Long productId);
 
 
 }
